replaced the appendToBody directive with prop
