+++ src/pointerMixin.js
@@ -34,6 +34,7 @@ export default {
     },
     isOpen () {
       this.pointerDirty = false
+      if (this.appendToBody) this.appendOptionsToBody()
     },
     pointer () {
       this.$refs.search.setAttribute('aria-activedescendant', this.id + '-' + this.pointer.toString())
